LayoutTests/platform/mac-tiger/http/tests/xmlhttprequest/failed-auth-expected.txt

Test for bug 13075: XMLHttpRequest with failed authentication should set status to 401.

Sync, no credentials: OK
Sync, incorrect credentials: OK
Async, no credentials: 0 != 401
Async, incorrect credentials: 0 != 401
